+++ example/lib/main.dart
@@ -1,9 +1,12 @@
-import 'dart:convert';
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 
 import 'package:flutter/material.dart';
-import 'package:okra_widget/okra_widget.dart';
+import 'package:okra_widget_official/okra_widget.dart';
 
-void main() => runApp(MyApp());
+Future main() async {
+  await dotenv.load(fileName: ".env");
+  runApp(MyApp());
+}
 
 class MyApp extends StatelessWidget {
   // This widget is the root of your application.
@@ -20,7 +23,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
@@ -30,8 +33,96 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
 
+  void buildWithOptions() {
+    var banks = [
+      "ecobank-nigeria",
+      "fidelity-bank",
+      "first-bank-of-nigeria",
+      "first-city-monument-bank",
+      "guaranty-trust-bank",
+      "access-bank",
+      "unity-bank",
+      "alat",
+      "polaris-bank",
+      "stanbic-ibtc-bank",
+      "standard-chartered-bank",
+      "sterling-bank",
+      "union-bank-of-nigeria",
+      "united-bank-for-africa",
+      "wema-bank",
+      "rubies-bank",
+      "kuda-bank"
+    ];
+
+    Okra.buildWithOptions(context,
+        key: dotenv.env['key']!,
+        token: dotenv.env['token']!,
+        color: "#3AB795",
+        products: ['auth', 'identity', 'balance', 'transactions'],
+        chargeAmount: 100,
+        chargeNote: "testing",
+        chargeType: "one-time",
+        chargeCurrency: "NGN",
+        environment: "production",
+        clientName: "clientName",
+        customerBvn: dotenv.env['bvn']!,
+        logo:
+        "https://dash.okra.ng/static/media/okra-logo.514fd943.png",
+        limit: 3,
+        currency: "NGN",
+        isCorporate: false,
+        showBalance: true,
+        geoLocation: true,
+        payment: false,
+        connectMessage:
+        "Which account do you want to connect with?",
+        callbackUrl: "",
+        redirectUrl: "",
+        widgetSuccess:
+        "Your account was successfully linked to SwipeNG",
+        widgetFailed:
+        "An unknown error occurred, please try again.",
+        guarantors: {
+          "status": false,
+          "message": "Okra requires you to add guarantors",
+          "number": 3,
+        },
+        filters: {"industry_type": "all", "banks": banks},
+        onSuccess: (data) {
+          print("Success");
+          print(data);
+        }, onError: (message) {
+          print("error");
+          print(message);
+        }, onClose: (message) {
+          print("close");
+          print(message);
+        });
+  }
+
+  void buildWithShortUrl() {
+    Okra.buildWithShortUrl(
+        context,
+        shortUrl: dotenv.env['url']!,
+        onSuccess: (data) {
+          print("Success");
+          print(data);
+        },
+        onError: ( message) {
+          print("error");
+          print(message);
+        },
+        onClose: (message) {
+          print("close");
+          print(message);
+        }
+    );
+  }
   @override
   Widget build(BuildContext context) {
+    final ButtonStyle buttonStyle = ButtonStyle(
+      backgroundColor: MaterialStateProperty.all(Colors.green,)
+    );
     return Scaffold(
       appBar: AppBar(
         // Here we take the value from the MyHomePage object that was created by
@@ -43,71 +134,25 @@ class _MyHomePageState extends State<MyHomePage> {
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             Text('Click button to open Okra Widget'),
-            RaisedButton(
-                color: Colors.green,
+            SizedBox(height: 40),
+            ElevatedButton(
                 child: Text(
-                  "Click me",
+                  "Build With Options",
                   style: TextStyle(color: Colors.white),
                 ),
+                style: buttonStyle,
                 onPressed: () async {
-                  var banks = [
-                    "ecobank-nigeria",
-                    "fidelity-bank",
-                    "first-bank-of-nigeria",
-                    "first-city-monument-bank",
-                    "guaranty-trust-bank",
-                    "access-bank",
-                    "unity-bank",
-                    "alat",
-                    "polaris-bank",
-                    "stanbic-ibtc-bank",
-                    "standard-chartered-bank",
-                    "sterling-bank",
-                    "union-bank-of-nigeria",
-                    "united-bank-for-africa",
-                    "wema-bank",
-                    "rubies-bank",
-                    "kuda-bank"
-                  ];
-
-
-                  var okraOptions = {
-                    "key": "z",
-                    "token": "5da6358130a943486f33dced",
-                    "products": [
-                      "auth",
-                      "balance",
-                      "identity",
-                      "transactions"
-                    ],
-                    "environment": "production",
-                    "clientName": "Bassey",
-                    "color" : "#9013FE",
-                    "limit" : "3",
-                    "isCorporate" : false,
-                    "connectMessage" : "Which account do you want to connect with?",
-                    "callback_url" : "",
-                    "redirect_url" : "",
-                    "logo" : "https://dash.okra.ng/static/media/okra-logo.514fd943.png",
-                    "widget_success" : "Your account was successfully linked to SwipeNG",
-                    "widget_failed" : "An unknown error occurred, please try again.",
-                    "currency" : "NGN",
-                    "noPeriodic" : true,
-                    "exp" : "",
-                    "success_title" : "null",
-                    "success_message" : "null",
-                    "guarantors" : {
-                      "status": false,
-                      "message": "Okra requires you to add guarantors",
-                      "number": 3,
-                    },
-                    "filter" : {
-                      "industry_type": "all",
-                      "banks": banks
-                    }
-                  };
-
-                  OkraHandler reply = await Okra.create(context, okraOptions);
+                  buildWithOptions();
+                }),
+            SizedBox(height: 20),
+            ElevatedButton(
+                child: Text(
+                  "Build With ShortUrl",
+                  style: TextStyle(color: Colors.white),
+                ),
+                style: buttonStyle,
+                onPressed: () async {
+                  buildWithShortUrl();
                 }),
           ],
         ),